test: customize sql log formatter - 동적 변경 지원
