--- src/test/java/com/meoguri/linkocean/test/support/common/P6spyLogMessageFormatConfiguration.java
@@ -1,17 +1,0 @@
-package com.meoguri.linkocean.test.support.common;
-
-import javax.annotation.PostConstruct;
-
-import org.springframework.boot.test.context.TestConfiguration;
-
-import com.p6spy.engine.spy.P6SpyOptions;
-
-@TestConfiguration
-public class P6spyLogMessageFormatConfiguration {
-
-	@PostConstruct
-	public void setLogMessageFormat() {
-		P6SpyOptions.getActiveInstance().setLogMessageFormat(CustomP6spySqlFormat.class.getName());
-	}
-
-}
