+++ src/test/java/com/meoguri/linkocean/test/support/service/BaseServiceTest.java
@@ -7,6 +7,8 @@ import static java.util.stream.Collectors.*;
 import java.util.Arrays;
 import java.util.List;
 
+import javax.persistence.EntityManager;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
@@ -32,12 +34,15 @@ import com.meoguri.linkocean.domain.user.entity.vo.Email;
 import com.meoguri.linkocean.domain.user.entity.vo.OAuthType;
 import com.meoguri.linkocean.domain.user.service.UserService;
 import com.meoguri.linkocean.domain.user.service.dto.GetUserResult;
-import com.meoguri.linkocean.test.support.common.P6spyLogMessageFormatConfiguration;
+import com.meoguri.linkocean.test.support.logging.p6spy.P6spyLogMessageFormatConfiguration;
 
 @Import(P6spyLogMessageFormatConfiguration.class)
 @Transactional
 @SpringBootTest
 public class BaseServiceTest {
+
+	@Autowired
+	protected EntityManager em;
 
 	@Autowired
 	private UserService userService;
